Fix failing webpack builds
vscode / tsc did not complain about unused variables for describe / it (from mocha) but ts-loader did. Not sure why. Adding an import to the spec files fixed the issue, but it feels a bit hacky, added a note to the TODO to look into this further, probably worth raising an issue against ts-loader